[refactor] [#1] 렌더링에 필요한 데이터들은 폴더로 분리
- sidebarTheme컴포넌트에서 사용되는 데이터 분리
- featureType컴포넌트에서 사용되는 데이터 분리
