+++ src/components/Sidebar/SidebarContentFewer/SidebarContentThemePresenter.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import ThemeItemPresenter from './ThemeItemPresenter';
 import styled from '../../../utils/styles/styled';
+import data from '../../../utils/redering-data/sidebarThemeData';
 
 const ThemeWrapper = styled.div`
   padding: 50px 16px;
@@ -14,39 +15,6 @@ const Title = styled.h2`
 const List = styled.ul`
   margin-top: 25px;
 `;
-
-const data = [
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '표준',
-  },
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '실버',
-  },
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '레트로',
-  },
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '흑백',
-  },
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '밤',
-  },
-  {
-    src:
-      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
-    name: '가지색',
-  },
-];
 
 function SidebarContentThemePresenter(): React.ReactElement {
   const [checkedItemIndex, setCheckedItemIndex] = useState(0);
