+++ src/utils/redering-data/featureTypeData.ts
@@ -1,0 +1,12 @@
+const data = [
+  { key: 'poi', name: 'POI' },
+  { key: 'road', name: '도로' },
+  { key: 'building', name: '건물' },
+  { key: 'terrain', name: '지형' },
+  { key: 'housing', name: '단지' },
+  { key: 'contour', name: '등고' },
+  { key: 'transit', name: '교통' },
+  { key: 'water', name: '수계' },
+];
+
+export default data;
