+++ src/utils/redering-data/sidebarThemeData.ts
@@ -1,0 +1,34 @@
+const data = [
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '표준',
+  },
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '실버',
+  },
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '레트로',
+  },
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '흑백',
+  },
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '밤',
+  },
+  {
+    src:
+      'https://i.pinimg.com/originals/8f/6d/97/8f6d971f3d086edaf6cee773991abb27.jpg',
+    name: '가지색',
+  },
+];
+
+export default data;
